Natalia-update-report
Tried to put the gradient as a background and not only have a part of it seen on screen. A little changes on value page - have the vision big and bellow that the 4 values. Changes of fonts and colors so it matches better our style guide and to make use of Google Fonts